unify solve! for SinkhornSolver and SinkhornBarycenterSolver
