+++ src/OptimalTransport.jl
@@ -40,6 +40,7 @@ include("entropic/sinkhorn_epsscaling.jl")
 include("entropic/sinkhorn_unbalanced.jl")
 include("entropic/sinkhorn_barycenter.jl")
 include("entropic/sinkhorn_barycenter_gibbs.jl")
+include("entropic/sinkhorn_solve.jl")
 
 include("quadratic.jl")
 include("quadratic_newton.jl")
